Link to the C++ standard library
Cadical depends on the C++ standard library so we
need to link to it. This will often already be in
the linker search path and should just work.

If it doesn’t, allow the user to symlink the
library to /usr/local/lib so they don’t have to
pass additional flags through cargo to the linker.

On MacOS, we need to explicitly link against a
newer version which we can install with brew:

```
brew install gcc
ln -s /usr/local/Cellar/gcc/*/lib/gcc/*/libstdc++.a /usr/local/lib/
```